requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pydantic==2.8.2
pyyaml==6.0.2
python-dateutil==2.9.0.post0
